+++ lib/cucumber/formatter/html.rb
@@ -8,7 +8,7 @@ module Cucumber
       include Io
 
       def initialize(config)
-        @io = ensure_io(config.out_stream)
+        @io = ensure_io(config.out_stream, config.error_stream)
         @html_formatter = Cucumber::HTMLFormatter::Formatter.new(@io)
         @html_formatter.write_pre_message